+++ src/cdkGlobal/design/question/RegularCustomer.java
@@ -3,12 +3,12 @@ package cdkGlobal.design.question;
 public class RegularCustomer implements BillStrategy{
 
 	@Override
-	public float calculateBill(float amount) {
-		float discount =0;
+	public double calculateBill(double amount) {
+		double discount =0;
 		if(amount>=5000 && amount <10000) {
-			discount = amount *(10/100);
+			discount = amount *(10d/100d);
 		}else{
-			discount = amount *(20/100);
+			discount = amount *(20d/100d);
 		}
 		return amount - discount;
 	}
